# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OMXS30 Screener API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SessionImport {
  /** The raw text output from the screener */
  rawText: string;
}

export interface MarketWeather {
  omxsValue: number;
  perf5d: number;
  perf1m: number;
  perf3m: number;
  rsi: number;
  trendLabel: string;
}

export type CandidateCategory = typeof CandidateCategory[keyof typeof CandidateCategory];


export const CandidateCategory = {
  MOMENTUM: 'MOMENTUM',
  SQUEEZE: 'SQUEEZE',
  STUDS: 'STUDS',
  WEAKEST: 'WEAKEST',
} as const;

export type CandidateDirection = typeof CandidateDirection[keyof typeof CandidateDirection];


export const CandidateDirection = {
  LONG: 'LONG',
  SHORT: 'SHORT',
} as const;

export type CandidateOutcomeProperty = typeof CandidateOutcomeProperty[keyof typeof CandidateOutcomeProperty];


export const CandidateOutcomeProperty = {
  WIN: 'WIN',
  LOSS: 'LOSS',
  SKIP: 'SKIP',
  PENDING: 'PENDING',
} as const;

export interface Candidate {
  id: number;
  sessionId: number;
  category: CandidateCategory;
  rank: number;
  ticker: string;
  price: number;
  rs3m: number;
  perf1m: number;
  rsi: number;
  pctB: number;
  atr: number;
  volMultiplier: number;
  distFrom20dH: number;
  /** @nullable */
  gapWarning?: number | null;
  direction: CandidateDirection;
  entryPrice: number;
  stopPrice: number;
  targetPrice: number;
  rr: number;
  oneR: number;
  outcome: CandidateOutcomeProperty;
  /** @nullable */
  exitPrice?: number | null;
  /** @nullable */
  outcomeNotes?: string | null;
}

export type CandidateWithSession = Candidate & {
  /** YYYY-MM-DD */
  sessionDate: string;
};

export type CandidateOutcomeOutcome = typeof CandidateOutcomeOutcome[keyof typeof CandidateOutcomeOutcome];


export const CandidateOutcomeOutcome = {
  WIN: 'WIN',
  LOSS: 'LOSS',
  SKIP: 'SKIP',
  PENDING: 'PENDING',
} as const;

export interface CandidateOutcome {
  outcome: CandidateOutcomeOutcome;
  /** @nullable */
  exitPrice?: number | null;
  /** @nullable */
  outcomeNotes?: string | null;
}

export interface SessionSummary {
  id: number;
  /** YYYY-MM-DD */
  date: string;
  marketWeather: MarketWeather;
  momentumCount: number;
  squeezeCount: number;
  studsCount: number;
  weakestCount: number;
  totalCandidates: number;
  /** Number of candidates with a non-PENDING outcome */
  trackedCount: number;
}

export interface SessionDetail {
  id: number;
  /** YYYY-MM-DD */
  date: string;
  marketWeather: MarketWeather;
  candidates: Candidate[];
}

export type CategoryStatsCategory = typeof CategoryStatsCategory[keyof typeof CategoryStatsCategory];


export const CategoryStatsCategory = {
  MOMENTUM: 'MOMENTUM',
  SQUEEZE: 'SQUEEZE',
  STUDS: 'STUDS',
  WEAKEST: 'WEAKEST',
} as const;

export interface CategoryStats {
  category: CategoryStatsCategory;
  total: number;
  wins: number;
  losses: number;
  skips: number;
  pending: number;
  /**
     * Win rate as a fraction (0-1), null if no resolved trades
     * @nullable
     */
  winRate: number | null;
}

export interface DashboardSummary {
  totalSessions: number;
  totalCandidates: number;
  resolvedTrades: number;
  wins: number;
  losses: number;
  /** @nullable */
  overallWinRate: number | null;
  byCategory: CategoryStats[];
}

export type ListCandidatesParams = {
category?: ListCandidatesCategory;
outcome?: ListCandidatesOutcome;
direction?: ListCandidatesDirection;
};

export type ListCandidatesCategory = typeof ListCandidatesCategory[keyof typeof ListCandidatesCategory];


export const ListCandidatesCategory = {
  MOMENTUM: 'MOMENTUM',
  SQUEEZE: 'SQUEEZE',
  STUDS: 'STUDS',
  WEAKEST: 'WEAKEST',
} as const;

export type ListCandidatesOutcome = typeof ListCandidatesOutcome[keyof typeof ListCandidatesOutcome];


export const ListCandidatesOutcome = {
  WIN: 'WIN',
  LOSS: 'LOSS',
  SKIP: 'SKIP',
  PENDING: 'PENDING',
} as const;

export type ListCandidatesDirection = typeof ListCandidatesDirection[keyof typeof ListCandidatesDirection];


export const ListCandidatesDirection = {
  LONG: 'LONG',
  SHORT: 'SHORT',
} as const;