# lib/api-zod/src/generated/types/candidate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OMXS30 Screener API
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateCategory } from './candidateCategory';
import type { CandidateDirection } from './candidateDirection';
import type { CandidateOutcomeProperty } from './candidateOutcomeProperty';

export interface Candidate {
  id: number;
  sessionId: number;
  category: CandidateCategory;
  rank: number;
  ticker: string;
  price: number;
  rs3m: number;
  perf1m: number;
  rsi: number;
  pctB: number;
  atr: number;
  volMultiplier: number;
  distFrom20dH: number;
  /** @nullable */
  gapWarning?: number | null;
  direction: CandidateDirection;
  entryPrice: number;
  stopPrice: number;
  targetPrice: number;
  rr: number;
  oneR: number;
  outcome: CandidateOutcomeProperty;
  /** @nullable */
  exitPrice?: number | null;
  /** @nullable */
  outcomeNotes?: string | null;
}