# lib/api-zod/src/generated/types/candidateOutcome.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OMXS30 Screener API
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateOutcomeOutcome } from './candidateOutcomeOutcome';

export interface CandidateOutcome {
  outcome: CandidateOutcomeOutcome;
  /** @nullable */
  exitPrice?: number | null;
  /** @nullable */
  outcomeNotes?: string | null;
}