# lib/api-zod/src/generated/types/candidateWithSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OMXS30 Screener API
 * OpenAPI spec version: 0.1.0
 */
import type { Candidate } from './candidate';

export type CandidateWithSession = Candidate & {
  /** YYYY-MM-DD */
  sessionDate: string;
};