# lib/api-zod/src/generated/types/categoryStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OMXS30 Screener API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryStatsCategory } from './categoryStatsCategory';

export interface CategoryStats {
  category: CategoryStatsCategory;
  total: number;
  wins: number;
  losses: number;
  skips: number;
  pending: number;
  /**
     * Win rate as a fraction (0-1), null if no resolved trades
     * @nullable
     */
  winRate: number | null;
}