# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OMXS30 Screener API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryStats } from './categoryStats';

export interface DashboardSummary {
  totalSessions: number;
  totalCandidates: number;
  resolvedTrades: number;
  wins: number;
  losses: number;
  /** @nullable */
  overallWinRate: number | null;
  byCategory: CategoryStats[];
}