# lib/api-zod/src/generated/types/marketWeather.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OMXS30 Screener API
 * OpenAPI spec version: 0.1.0
 */

export interface MarketWeather {
  omxsValue: number;
  perf5d: number;
  perf1m: number;
  perf3m: number;
  rsi: number;
  trendLabel: string;
}