# lib/api-zod/src/generated/types/sessionDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OMXS30 Screener API
 * OpenAPI spec version: 0.1.0
 */
import type { Candidate } from './candidate';
import type { MarketWeather } from './marketWeather';

export interface SessionDetail {
  id: number;
  /** YYYY-MM-DD */
  date: string;
  marketWeather: MarketWeather;
  candidates: Candidate[];
}