# lib/api-zod/src/generated/types/sessionImport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OMXS30 Screener API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionImport {
  /** The raw text output from the screener */
  rawText: string;
}