# lib/api-zod/src/generated/types/sessionSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OMXS30 Screener API
 * OpenAPI spec version: 0.1.0
 */
import type { MarketWeather } from './marketWeather';

export interface SessionSummary {
  id: number;
  /** YYYY-MM-DD */
  date: string;
  marketWeather: MarketWeather;
  momentumCount: number;
  squeezeCount: number;
  studsCount: number;
  weakestCount: number;
  totalCandidates: number;
  /** Number of candidates with a non-PENDING outcome */
  trackedCount: number;
}